+++ lib/_stream_duplex.js
@@ -29,33 +29,15 @@
 module.exports = Duplex;
 
 const util = require('util');
-const Readable = require('_stream_readable');
-const Writable = require('_stream_writable');
+const DuplexBase = require('internal/streams/duplex_base');
 
-util.inherits(Duplex, Readable);
-
-{
-  // avoid scope creep, the keys array can then be collected
-  const keys = Object.keys(Writable.prototype);
-  for (var v = 0; v < keys.length; v++) {
-    const method = keys[v];
-    if (!Duplex.prototype[method])
-      Duplex.prototype[method] = Writable.prototype[method];
-  }
-}
+util.inherits(Duplex, DuplexBase);
 
 function Duplex(options) {
   if (!(this instanceof Duplex))
     return new Duplex(options);
 
-  Readable.call(this, options);
-  Writable.call(this, options);
-
-  if (options && options.readable === false)
-    this.readable = false;
-
-  if (options && options.writable === false)
-    this.writable = false;
+  DuplexBase.call(this, options);
 
   this.allowHalfOpen = true;
   if (options && options.allowHalfOpen === false) {
